+++ src/constants/apiRoutes.js
@@ -1,0 +1,2 @@
+export const DOCUMENTS_URL = "https://app-sandbox.mifiel.com/api/v1/documents";
+export const USER_ME_URL = "https://app-sandbox.mifiel.com/api/v1/users/me";