+++ src/pages/components/Avatar.jsx
@@ -8,7 +8,7 @@ const Avatar = ({ className }) => {
 
   return (
     <div
-      className={`${className} rounded-full w-10 h-10 flex justify-center items-center text-2xl`}
+      className={`${className} rounded-full w-10 h-10 flex justify-center items-center text-xl`}
     >
       {avatar}
     </div>